Reorganize existing boilerplate models
