+++ boilerplate/src/app/root-component.tsx
@@ -2,7 +2,7 @@ import "../i18n"
 import * as React from "react"
 import { setupRootStore } from "./setup-root-store"
 import { StatefulNavigator } from "../navigation"
-import { RootStore } from "../models/root-store"
+import { RootStore } from "../app/root-store"
 import { Provider } from "mobx-react"
 import { BackButtonHandler } from "../navigation/back-button-handler"
 import { contains } from "ramda"
