+++ boilerplate/src/app/setup-root-store.ts
@@ -1,6 +1,6 @@
 import { onSnapshot } from "mobx-state-tree"
-import { RootStoreModel, RootStore } from "../models/root-store"
-import { Environment } from "../models/environment"
+import { RootStoreModel, RootStore } from "./root-store"
+import { Environment } from "./environment"
 import * as storage from "../lib/storage"
 import { Reactotron } from "../services/reactotron"
 import { Api } from "../services/api"
