--- boilerplate/src/models/root-store/index.ts
@@ -1,1 +1,0 @@
-export * from "./root-store"
