+++ boilerplate/src/navigation/navigation-events.ts
