+++ boilerplate/src/navigation/stateful-navigator.tsx
@@ -2,7 +2,7 @@ import * as React from "react"
 import { inject, observer } from "mobx-react"
 import { addNavigationHelpers } from "react-navigation"
 import { RootNavigator } from "./root-navigator"
-import { NavigationStore } from "../models/navigation-store/navigation-store"
+import { NavigationStore } from "../navigation/navigation-store"
 import throttle from "lodash.throttle"
 
 interface StatefulNavigatorProps {
